refactor: move roundRect polyfill to separate module and update renderer imports
